Migrate to dedicated Twig namespace
